Add files via upload
these are really not my proudest coding moments, but it is what it is. 

p1: i solved it with nested loops: the intention is to loop through each element inside a 2D matrix representing the "engine". each time an element is a digit, i proceed to to find a cluster up til the nearest symbol. i then do a flagged check for if this cluster has a symbol surrounding it - albeit letter by letter

p2: i took a different approach. i found "*", before finding nearby digits in the positions surrounding its element. i then spread out from that digit to find clusters, and then proceed to attach all clusters inside a separate list. if the number of clusters matches the requirements, i proceed to calculate what is needed.

all in all - i hate loops.